# apps/web/openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "AuthenticationErrorResponse": {
        "anyOf": [
          {
            "$ref": "#/components/schemas/UnauthorizedErrorResponse"
          },
          {
            "$ref": "#/components/schemas/SessionInvalidErrorResponse"
          }
        ],
        "description": "Bearer 缺失、无效或会话失效。",
        "title": "AuthenticationErrorResponse"
      },
...
      "CharacterListResponse": {
        "properties": {
          "code": {
            "const": 0,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "$ref": "#/components/schemas/CharacterListData"
          },
          "msg": {
            "const": "SUCCESS",
            "title": "Msg",
            "type": "string"
          }
        },
        "required": [
          "code",
          "msg",
          "data"
        ],
        "title": "CharacterListResponse",
        "type": "object"
      },
...
      "HTTPErrorResponse": {
        "anyOf": [
          {
            "$ref": "#/components/schemas/InternalErrorResponse"
          },
          {
            "$ref": "#/components/schemas/ValidationErrorResponse"
          },
          {
            "$ref": "#/components/schemas/UnauthorizedErrorResponse"
          },
          {
            "$ref": "#/components/schemas/ForbiddenErrorResponse"
          },
          {
            "$ref": "#/components/schemas/NotFoundErrorResponse"
          },
          {
            "$ref": "#/components/schemas/ConflictErrorResponse"
          },
          {
            "$ref": "#/components/schemas/RateLimitedErrorResponse"
          },
          {
            "$ref": "#/components/schemas/ServiceUnavailableErrorResponse"
          },
          {
            "$ref": "#/components/schemas/RegistrationDisabledErrorResponse"
          },
          {
            "$ref": "#/components/schemas/AccountIdentifierUnavailableErrorResponse"
          },
          {
            "$ref": "#/components/schemas/RegistrationRateLimitedErrorResponse"
          },
          {
            "$ref": "#/components/schemas/InvalidCredentialsErrorResponse"
          },
          {
            "$ref": "#/components/schemas/LoginRateLimitedErrorResponse"
          },
          {
            "$ref": "#/components/schemas/SessionInvalidErrorResponse"
          },
          {
            "$ref": "#/components/schemas/CurrentPasswordInvalidErrorResponse"
          },
          {
            "$ref": "#/components/schemas/MediaInvalidErrorResponse"
          },
          {
            "$ref": "#/components/schemas/MediaTooLargeErrorResponse"
          }
        ],
        "description": "所有稳定错误响应的联合类型，用于未预见状态的兜底契约。",
        "title": "HTTPErrorResponse"
      },
...
      "NotFoundErrorResponse": {
        "properties": {
          "code": {
            "const": 10004,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "additionalProperties": true,
            "title": "Data",
            "type": "object"
          },
          "msg": {
            "const": "NOT_FOUND",
            "title": "Msg",
            "type": "string"
          }
        },
        "required": [
          "data",
          "code",
          "msg"
        ],
        "title": "NotFoundErrorResponse",
        "type": "object"
      },
...
      "ProjectDetailResponse": {
        "properties": {
          "code": {
            "const": 0,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "$ref": "#/components/schemas/ProjectDetailData"
          },
          "msg": {
            "const": "SUCCESS",
            "title": "Msg",
            "type": "string"
          }
        },
        "required": [
          "code",
          "msg",
          "data"
        ],
        "title": "ProjectDetailResponse",
        "type": "object"
      },
...
      "ServiceUnavailableErrorResponse": {
        "properties": {
          "code": {
            "const": 10007,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "additionalProperties": true,
            "title": "Data",
            "type": "object"
          },
          "msg": {
            "const": "SERVICE_UNAVAILABLE",
            "title": "Msg",
            "type": "string"
          }
        },
        "required": [
          "data",
          "code",
          "msg"
        ],
        "title": "ServiceUnavailableErrorResponse",
        "type": "object"
      },
...
      "ValidationErrorResponse": {
        "properties": {
          "code": {
            "const": 10001,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "additionalProperties": true,
            "title": "Data",
            "type": "object"
          },
          "msg": {
            "const": "VALIDATION_ERROR",
            "title": "Msg",
            "type": "string"
          }
        },
        "required": [
          "data",
          "code",
          "msg"
        ],
        "title": "ValidationErrorResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "api",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/projects/{project_id}": {
      "get": {
        "operationId": "getProject",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthenticationErrorResponse"
                }
              }
            },
            "description": "Unauthorized",
            "headers": {
              "WWW-Authenticate": {
                "description": "客户端应使用 Bearer 认证方案",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceUnavailableErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Project Detail",
        "tags": [
          "projects"
        ]
      }
    },
    "/api/v1/projects/{project_id}/characters": {
      "get": {
        "operationId": "listProjectCharacters",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthenticationErrorResponse"
                }
              }
            },
            "description": "Unauthorized",
            "headers": {
              "WWW-Authenticate": {
                "description": "客户端应使用 Bearer 认证方案",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceUnavailableErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Characters",
        "tags": [
          "planning"
        ]
      },
      "post": {
        "operationId": "createProjectCharacter",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CharacterCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthenticationErrorResponse"
                }
              }
            },
            "description": "Unauthorized",
            "headers": {
              "WWW-Authenticate": {
                "description": "客户端应使用 Bearer 认证方案",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConflictErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceUnavailableErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Post Character",
        "tags": [
          "planning"
        ]
      }
    },
    "/api/v1/projects/{project_id}/characters/reorder": {
      "post": {
        "operationId": "reorderProjectCharacters",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CharacterReorderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthenticationErrorResponse"
                }
              }
            },
            "description": "Unauthorized",
            "headers": {
              "WWW-Authenticate": {
                "description": "客户端应使用 Bearer 认证方案",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConflictErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceUnavailableErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Post Character Reorder",
        "tags": [
          "planning"
        ]
      }
    },
    "/api/v1/projects/{project_id}/characters/{character_id}": {
      "delete": {
        "operationId": "deleteProjectCharacter",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "character_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Character Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResourceDeleteResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthenticationErrorResponse"
                }
              }
            },
            "description": "Unauthorized",
            "headers": {
              "WWW-Authenticate": {
                "description": "客户端应使用 Bearer 认证方案",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConflictErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceUnavailableErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Character Endpoint",
        "tags": [
          "planning"
        ]
      },
      "patch": {
        "operationId": "updateProjectCharacter",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "character_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Character Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CharacterUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthenticationErrorResponse"
                }
              }
            },
            "description": "Unauthorized",
            "headers": {
              "WWW-Authenticate": {
                "description": "客户端应使用 Bearer 认证方案",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConflictErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceUnavailableErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Patch Character",
        "tags": [
          "planning"
        ]
      }
    },
...
    "/api/v1/projects/{project_id}/export": {
      "get": {
        "operationId": "exportProject",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "format",
            "required": false,
            "schema": {
              "default": "markdown",
              "enum": [
                "markdown",
                "plain_text"
              ],
              "title": "Format",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/markdown": {
                "schema": {
                  "type": "string"
                }
              },
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "UTF-8 作品正文导出文件"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthenticationErrorResponse"
                }
              }
            },
            "description": "Unauthorized",
            "headers": {
              "WWW-Authenticate": {
                "description": "客户端应使用 Bearer 认证方案",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceUnavailableErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Export",
        "tags": [
          "planning"
        ]
      }
    },
...
  }
}